+++ EssentialFeed/EssentialFeed/Image Comments API/ImageCommentsMapper.swift
@@ -5,7 +5,7 @@
 
 import Foundation
 
-enum ImageCommentsMapper {
+public enum ImageCommentsMapper {
     private struct Root: Decodable {
         private let items: [Item]
         
@@ -25,7 +25,7 @@ enum ImageCommentsMapper {
         }
     }
     
-    static func map(_ data: Data, from response: HTTPURLResponse) throws -> [ImageComment] {
+    public static func map(_ data: Data, from response: HTTPURLResponse) throws -> [ImageComment] {
         let decoder = JSONDecoder()
         decoder.dateDecodingStrategy = .iso8601
